all type of double to float

+++ include/HitRecord.hpp
@@ -6,8 +6,8 @@
 class HitRecord{
 public:
 	int _surface;	// which surface (surface id)
-	double _t;	// where hit
-	Point3d _hit_point;
+	float _t;	// where hit
+	Point3f _hit_point;
 };
 
 #endif
